requirements.txt
asgiref==3.2.10
Django==3.0
django-filter==2.3.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
